chore: reorganize composition wiring into feature/shared modules and update imports
