+++ src/composition/shared/http/middleware.ts
@@ -1,5 +1,5 @@
 import { flattenViolations } from '@koala-ts/framework/validator';
 import { validationMiddleware } from '@/app/shared/interface/http';
-import { validate } from '@/composition/http/validation';
+import { validate } from '@/composition/shared/http/validation';
 
 export const validateRequest = validationMiddleware(validate, flattenViolations);
