+++ src/controller/AuthController.ts
@@ -1,8 +1,8 @@
 import { type HttpScope, Route } from '@koala-ts/framework';
 import { LoginRequest } from '@/app/authentication/interface/http/login-request';
 import { loginResponse } from '@/app/authentication/interface/http/login-response';
-import { loginUser, validateLoginRequest } from '@/composition/authentication/login';
-import { mapResultToHttp } from '@/composition/http/normalization';
+import { loginUser, validateLoginRequest } from '@/composition/auth/login';
+import { mapResultToHttp } from '@/composition/shared/http/normalization';
 
 export class AuthController {
   @Route({ method: 'POST', path: '/api/v1/login', middleware: [validateLoginRequest] })
